Avoid Throwable#addSuppressed, use CompositeException
I would say (without backing data) that it's fairly common in the
ecosystem to sometimes use 'case object' when defining a throwable.
This makes addSuppressed a memory leak risk.  There is the option to
disable suppressed exceptions, which makes addSuppressed a noop, but
that just makes it swallow errors then...

So for now I think it might be safest to just use CompositeException
(like we were already on JS).  Though I do share Ross' concern about
root causes...