Do some doc fixing
This started out as me just attempting to provide a link to the
"Acquiring and releasing `Resource`s" section of the tutorial from with
tin the `Resource` data type doc. But then I went down a bit of a rabbit
hole...

There was a Scala 2.12.6 vs 2.12.7 mismatch between `build.sbt` and
`.travis.yml`. This meant that the build hasn't actually been compiling
the docs recently. So I fixed a few spots that weren't compiling in the
docs (though I'm not sure whether I fixed them in the preferred way). I
went ahead and attempted to make the scala version check for the docs
build a bit more lenient to hopefully avoid a similar issue in the
future. Admittedly my bash skills are lacking, so we should probably
check that it works as expected in the Travis build.