Update the fiber yielding explanation in basics.md
> we can decide when we yield to other threads from the same pool by calling shift

> Allowing different threads to advance their work is called fairness

These are a bit misleading. We actually yield to other pending *fibers*, not other threads. Also, fairness is allowing different *fibers* to advance their work, not different threads (at least in this context... fairness among threads is also a thing, but that's achieved on CPU level).

It might sound like nitpicking, but to me current version sounds as if, when `shift` is called, some or all of the busy threads might stop their work and replace their current fibers with some of the pending fibers. And that will not happen. When we `shift` some fiber F on some thread T, that means that T might get another fiber, and that F might continue running on a different thread (if any are available). But busy threads will not be disrupted.