Fixes #2606
In the case of IOApp, the main fiber is registered as an external fiber
but then also monitored when it is scheduled on the workpool.

Whilst somewhat hacky (I can't envision another situation in which this
arises), removing any fibers known to be suspended from the set of
foreign fibers seems reasonable at least.