Simplify ContState
Instead of having 4 separate states (initial, waiting,
winner and result) now we only have 3: initial, waiting
and result. Instead of an `AtomicInteger` storing tags,
now we use an `AtomicReference` for the result, and 2
sentinels: `null` for "initial", and a private object
for "waiting".

This allows us to publish the result with a single
CAS (instead of a CAS and a volatile write). We can
also get rid of the spinning in `Get` (since we don't
have to wait for the transition between "winner" and
"result").

There should be no observable change in behavior
(except maybe a small performance improvement).