Seal - Simplified Deferred of Unit.
Concurrency libraries, such as those based on `cats-effect` such as
FS2, often have this pattern in which a lightweight thread / fiber
has to await (and be parked while waiting) on a condition, while
other fiber will mark that the wait is over.

Notably, this is a synchronisation or control-flow problem, as
it only relates to the ordering and sequencing between the
processes performed by associated fibers. It is not a data-flow
or communications problem, in which a fiber needs to wait until
it can get a piece of data that is being computed by another one.

In cats-effect based libraries, such as FS2, a `Deferred[Unit]` is
often use for this problem, but Deferred is a data-flow solution,
not a control-flow solution.

This PR introduces a specialised variant of Deferred, a `Seal`,
that is intended.
The word "Seal" is used in the sense of letter seals: all that
matters is whether they are broken or not, they start unbroken,
and they can only be broken once.